moved customisation of html reports to fixtures.py from conftest.py, so as to maintain the control in vane framework
