+++ sample_network_tests/conftest.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 #
-# Copyright (c) 2019, Arista Networks EOS+
+# Copyright (c) 2023, Arista Networks EOS+
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -28,14 +28,11 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
-# pylint: disable=import-error,no-name-in-module,invalid-name
+# pylint: disable=invalid-name
 
 """ Alter behavior of PyTest """
 
-import re
 import pytest
-from py.xml import html
-from vane.config import test_defs
 
 pytest_plugins = "vane.fixtures"
 
@@ -68,88 +65,3 @@ def definitions(request):
 
     cli_arg = request.config.getoption("--definitions")
     return cli_arg
-
-
-def find_nodeid(nodeid):
-    """Return device parameter
-
-    Args:
-        nodeid (str): Device Name
-
-    Return: Name of device
-
-    """
-
-    if re.match(r".*\[(.*)\]", nodeid):
-        return re.match(r".*\[(.*)\]", nodeid)[1]
-
-    return "NONE"
-
-
-def read_test_id(report):
-    """Returns test id for the test case being executed
-
-    Args:
-        report: pytest report
-    """
-    last_double_colon_index = report.nodeid.rfind("::")
-    last_open_bracket_index = report.nodeid.rfind("[")
-    test_case_name = report.nodeid[
-        last_double_colon_index + 2 : last_open_bracket_index  # noqa: E203
-    ]
-
-    for test_suite in test_defs["test_suites"]:
-        for test_case in test_suite["testcases"]:
-            if test_case_name == test_case["name"]:
-                if "test_id" in test_case:
-                    return test_case["test_id"]
-                return ""
-
-    return ""
-
-
-def pytest_html_results_table_header(cells):
-    """Create custom PyTest-HTML Header Row
-
-    Args:
-        cells: Cell data
-    """
-
-    cells.insert(2, html.th("Description"))
-    cells.insert(1, html.th("Test ID"))
-    cells.insert(1, html.th("Device", class_="sortable string", col="device"))
-    cells.pop()
-
-
-def pytest_html_results_table_row(report, cells):
-    """Create custom PyTest-HTML report row
-
-    Args:
-        report: pytest report
-        cells: Cell data
-    """
-
-    cells.insert(2, html.td(getattr(report, "description", "")))
-    cells.insert(1, html.td(getattr(report, "test_id", "")))
-    cells.insert(1, html.td(find_nodeid(report.nodeid), class_="col-device"))
-    cells.pop()
-
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_runtest_makereport(item):
-    """Called to create a _pytest.reports.TestReport for each of the setup,
-    call and teardown runtest phases of a test item.
-
-    """
-
-    outcome = yield
-    report = outcome.get_result()
-
-    report.test_id = read_test_id(report)
-
-    if str(item.function.__doc__).split("Args:", maxsplit=1)[0]:
-        report.description = str(item.function.__doc__).split("Args:", maxsplit=1)[0]
-    elif str(item.function.__doc__):
-        report.description = str(item.function.__doc__)
-    else:
-        report.description = "No Description"
